second

+++ src/app/addcart/addcart.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AddtocartService } from '../services/addtocart.service';
+import { PaymentService } from '../services/payment.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-addcart',
@@ -9,60 +11,82 @@ import { AddtocartService } from '../services/addtocart.service';
 export class AddcartComponent implements OnInit {
   cartitems: any;
   total:number = 0;
+  quantitycountZer0: number;
+  quantitycountMax: number;
 
-  constructor(private addcartservice:AddtocartService) { }
+  constructor(private addcartservice:AddtocartService,
+    private paymentservice:PaymentService,
+    private router:Router) { }
 
   ngOnInit() {
     this.addcartservice.getCart(localStorage.getItem('user_id'))
     .subscribe((result_cartitem)=>{
-      console.log(result_cartitem.cartitem);
       this.cartitems=result_cartitem.cartitem;
+      this.totalCartPrice();
     },
     (err)=>{
       console.log(err);
     })
   }
 
+  totalCartPrice(){
+    for(let i=0;i<this.cartitems.length;i++){
+      this.total += this.cartitems[i].cartprice;
+    }
+  }
 
   totalPrice(){
     this.total = 0;
-    for(var i=0;i<this.cartitems.length;i++){
+    for(let i=0;i<this.cartitems.length;i++){
       this.total += (this.cartitems[i]._id.food_price * this.cartitems[i].quantitycount);
     }
   }
 
+  
   add(pid){
-    console.log(pid);
-    for(var i=0;i<this.cartitems.length;i++){
+    for(let i=0;i<this.cartitems.length;i++){
       if(this.cartitems[i]._id.food_id === pid)
       {  
-        this.cartitems[i].quantitycount += 1;
-        this.addcartservice.addcartItem(this.cartitems[i],this.cartitems[i]._id.food_id)
-        .subscribe((result_added)=>{
-          console.log(result_added);
-        })
+        if(this.cartitems[i].quantitycount==5){
+          this.quantitycountMax=1;
+        }
+        else{
+          this.quantitycountZer0=0;
+          this.cartitems[i].quantitycount += 1;
+          this.addcartservice.addcartItem(this.cartitems[i],this.cartitems[i]._id.food_id)
+          .subscribe((result_added)=>{
+            console.log("Successfully added");
+          })
+        }
       }           
     }
     this.totalPrice();
-    console.log(this.cartitems);
   }
 
   del(pid){
-    console.log(pid);
-    for(var i=0;i<this.cartitems.length;i++){
+    for(let i=0;i<this.cartitems.length;i++){
       if(this.cartitems[i]._id.food_id === pid)
       {  
-        this.cartitems[i].quantitycount -= 1;
-        this.addcartservice.deletecartItem(this.cartitems[i]._id.food_id,localStorage.getItem('user_id'))
-        .subscribe((result_deleted)=>{
-          console.log(result_deleted);
-        })
+        if(this.cartitems[i].quantitycount==0){
+          this.quantitycountZer0=1;
+        }
+        else{
+          this.cartitems[i].quantitycount -= 1;
+          this.quantitycountMax=0;
+          this.addcartservice.deletecartItem(this.cartitems[i]._id.food_id,localStorage.getItem('user_id'))
+          .subscribe((result_deleted)=>{
+            console.log("Successfully deleted");
+          })
+        }
       }           
     }
-    
-
     this.totalPrice();
-    console.log(this.cartitems);
+  }
+
+  proceedPay(total){
+    // this.paymentservice.cartItemTotal.emit(total);
+    this.paymentservice.cartTotalPrice.next(total);
+    this.router.navigateByUrl('/pay');
   }
 
 }
